+++ CHANGELOG.md
@@ -1,2 +1,7 @@
 1.0.20
 - Initial Release
+
+1.0.21
+- Bug in custom font hd44780
+- Add column properties for free writepos
+- Add Node for big font